feat: added functionality to show a detailed lecture if fetch is done
This function will check whether the selected lecture has been pre-fetched. If that is the case, then the activity is going to be switched to the lecture-fragment activity, after a short delay. In case the details are not yet fetched a loading bar at the bottom (SnackBar) will show the passed time.